+++ packages/amplify-category-function/src/provider-utils/awscloudformation/utils/getDependentFunction.ts
@@ -1,24 +1,25 @@
 import { $TSContext, $TSObject } from 'amplify-cli-core';
 import * as path from 'path';
-import { loadFunctionParameters } from './loadFunctionParameters';
+import { loadFunctionParameters, loadFunctionStackAsJSON } from './loadFunctionParameters';
 import { ServiceName } from './constants';
 import { categoryName } from '../../../constants';
 
-export async function lambdasWithApiDependency(
+export async function lambdasWithMissingApiDependency(
   context: $TSContext,
   allResources: $TSObject[],
   backendDir: string,
-  modelsDeleted: string[],
+  existingModels: string[],
 ) {
-  //get the List of functions dependent on deleted model
+  //get the List of functions dependent on deleted models
   let dependentFunctions = [];
-  const lambdaFuncResources = allResources.filter(
-    resource =>
+  const lambdaFuncResources = allResources.filter(resource => {
+    return (
       resource.service === ServiceName.LambdaFunction &&
       resource.mobileHubMigrated !== true &&
       resource.dependsOn !== undefined &&
-      resource.dependsOn.find(val => val.category === 'api'),
-  );
+      resource.dependsOn.find(val => val.category === 'api')
+    );
+  });
 
   // initialize function parameters for update
   for (const lambda of lambdaFuncResources) {
@@ -29,9 +30,22 @@ export async function lambdasWithApiDependency(
 
     if (typeof selectedCategories === 'object' && selectedCategories !== null) {
       for (const selectedResources of Object.values(selectedCategories)) {
-        deletedModelFound = Object.keys(selectedResources).some(r => modelsDeleted.includes(r));
+        deletedModelFound = Object.keys(selectedResources).some(r => !existingModels.includes(r));
         if (deletedModelFound) {
           dependentFunctions.push(lambda);
+        }
+      }
+    } else {
+      // In the case that the function-parameters.json does not have info on dependencies, we check the CFN stack.
+      // If the dependency is only a dynamo stream trigger, we don't have info in function-parameters.json to work with
+      const lambdaStackJson = loadFunctionStackAsJSON(resourceDirPath, lambda.resourceName);
+      if (lambdaStackJson?.Resources) {
+        const triggerPolicyTables = Object.keys(lambdaStackJson.Resources).filter(key => key.startsWith('LambdaTriggerPolicy'));
+        for (const triggerPolicy of triggerPolicyTables) {
+          const tableName = triggerPolicy.match(/(?<=LambdaTriggerPolicy)(.*)/g)?.[0];
+          if (tableName && !existingModels.includes(tableName)) {
+            dependentFunctions.push(lambda);
+          }
         }
       }
     }