+++ packages/amplify-category-function/src/provider-utils/awscloudformation/utils/updateDependentFunctionCfn.ts
@@ -7,11 +7,11 @@ import * as path from 'path';
 import { functionParametersFileName } from './constants';
 import { categoryName } from '../../../constants';
 
-export async function updateDependentFunctionsCfn(
+export async function updateMissingDependencyFunctionsCfn(
   context: $TSContext,
   dependentFunctionResource: $TSObject[],
   backendDir: string,
-  modelsDeleted: string[],
+  existingModels: string[],
   apiResource: string,
 ) {
   // remove function parameters from if there is dependency
@@ -42,26 +42,28 @@ export async function updateDependentFunctionsCfn(
       },
     };
 
-    for (const selectedCategory of Object.keys(selectedCategories)) {
-      // update function parameters resouce parameters with deleted models data
-      // remove the deleted @model
-      const selectedResources = selectedCategories[selectedCategory];
-      for (const resourceName of Object.keys(selectedResources)) {
-        if (!modelsDeleted.includes(resourceName)) {
-          const resourcePolicy = selectedResources[resourceName];
-          const { permissionPolicies, cfnResources } = await getResourcesForCfn(
-            context,
-            resourceName,
-            resourcePolicy,
-            apiResource,
-            selectedCategory,
-          );
-          categoryPolicies = categoryPolicies.concat(permissionPolicies);
-          if (!permissions[selectedCategory]) {
-            permissions[selectedCategory] = {};
+    if (selectedCategories) {
+      for (const selectedCategory of Object.keys(selectedCategories)) {
+        // update function parameters resouce parameters with deleted models data
+        // remove the deleted @model
+        const selectedResources = selectedCategories[selectedCategory];
+        for (const resourceName of Object.keys(selectedResources)) {
+          if (existingModels.includes(resourceName)) {
+            const resourcePolicy = selectedResources[resourceName];
+            const { permissionPolicies, cfnResources } = await getResourcesForCfn(
+              context,
+              resourceName,
+              resourcePolicy,
+              apiResource,
+              selectedCategory,
+            );
+            categoryPolicies = categoryPolicies.concat(permissionPolicies);
+            if (!permissions[selectedCategory]) {
+              permissions[selectedCategory] = {};
+            }
+            permissions[selectedCategory][resourceName] = resourcePolicy;
+            resources = resources.concat(cfnResources);
           }
-          permissions[selectedCategory][resourceName] = resourcePolicy;
-          resources = resources.concat(cfnResources);
         }
       }
     }
@@ -72,6 +74,8 @@ export async function updateDependentFunctionsCfn(
     functionParameters.dependsOn = dependsOn;
     functionParameters.lambdaLayers = currentParameters.lambdaLayers;
     updateCFNFileForResourcePermissions(resourceDirPath, functionParameters, currentParameters, apiResource);
+    // In some cases, we need to ensure that API Dynamo event sources are removed when the model doesn't exist
+    removeDeletedApiDynamoEventSourcesFromCfn(resourceDirPath, lambda.resourceName, existingModels);
     // assign new permissions to current permissions to update function-parameters file
     currentParameters.permissions = permissions;
     // update function-parameters file
@@ -84,9 +88,7 @@ export async function updateDependentFunctionsCfn(
   }
 }
 
-export function addAppSyncInvokeMethodPermission(
-  functionName: string,
-) {
+export function addAppSyncInvokeMethodPermission(functionName: string) {
   const resourceDirPath = pathManager.getResourceDirectoryPath(undefined, categoryName, functionName);
   const cfnFileName = `${functionName}-cloudformation-template.json`;
   const cfnFilePath = path.join(resourceDirPath, cfnFileName);
@@ -97,12 +99,43 @@ export function addAppSyncInvokeMethodPermission(
       Type: 'AWS::Lambda::Permission',
       Properties: {
         FunctionName: {
-          Ref: "LambdaFunction"
+          Ref: 'LambdaFunction',
         },
-        Action: "lambda:InvokeFunction",
-        Principal: "appsync.amazonaws.com",
+        Action: 'lambda:InvokeFunction',
+        Principal: 'appsync.amazonaws.com',
       },
     };
   }
   JSONUtilities.writeJson(cfnFilePath, cfnContent);
+}
+
+function removeDeletedApiDynamoEventSourcesFromCfn(resourceDirPath: string, resourceName: string, existingModels: string[]) {
+  const sanitizedExistingModels = existingModels.map(value => {
+    return value.slice(0, value.indexOf(':'));
+  });
+  const cfnFilePath = path.join(resourceDirPath, `${resourceName}-cloudformation-template.json`);
+  const functionCfn = JSONUtilities.readJson<$TSAny>(cfnFilePath);
+  let tableDeletions: string[] = [];
+  if (functionCfn?.Resources) {
+    const triggerPolicyTables = Object.keys(functionCfn.Resources).filter(key => key.startsWith('LambdaTriggerPolicy'));
+    for (let triggerPolicy of triggerPolicyTables) {
+      const tableName = triggerPolicy.match(/(?<=LambdaTriggerPolicy)(.*)/g)?.[0];
+      if (tableName && !sanitizedExistingModels.includes(tableName)) {
+        const triggerPolicyStatement: $TSAny = functionCfn?.Resources?.[triggerPolicy]?.Properties?.PolicyDocument?.Statement;
+        for (const statement of triggerPolicyStatement ?? []) {
+          const apiIDCheckString: string = statement?.Resource?.['Fn::ImportValue']?.['Fn::Sub'];
+          if (apiIDCheckString && apiIDCheckString.includes('GraphQLAPIId')) {
+            tableDeletions.push(tableName);
+            break;
+          }
+        }
+      }
+    }
+  }
+
+  for (let tableName of tableDeletions) {
+    delete functionCfn?.Resources?.[`LambdaTriggerPolicy${tableName}`];
+    delete functionCfn?.Resources?.[`LambdaEventSourceMapping${tableName}`];
+  }
+  JSONUtilities.writeJson(cfnFilePath, functionCfn);
 }