+++ packages/amplify-dotnet-function-template-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.24](https://github.com/aws-amplify/amplify-cli/compare/amplify-dotnet-function-template-provider@2.2.23...amplify-dotnet-function-template-provider@2.2.24) (2022-02-25)
+
+**Note:** Version bump only for package amplify-dotnet-function-template-provider
+
+
+
+
+
 ## [2.2.23](https://github.com/aws-amplify/amplify-cli/compare/amplify-dotnet-function-template-provider@2.2.22...amplify-dotnet-function-template-provider@2.2.23) (2022-02-15)
 
 **Note:** Version bump only for package amplify-dotnet-function-template-provider