+++ packages/graphql-key-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.24](https://github.com/aws-amplify/amplify-cli/compare/graphql-key-transformer@3.2.23...graphql-key-transformer@3.2.24) (2022-02-25)
+
+**Note:** Version bump only for package graphql-key-transformer
+
+
+
+
+
 ## [3.2.23](https://github.com/aws-amplify/amplify-cli/compare/graphql-key-transformer@3.2.22...graphql-key-transformer@3.2.23) (2022-02-15)
 
 **Note:** Version bump only for package graphql-key-transformer