+++ packages/graphql-transformer-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.4.7](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformer-core@7.4.6...graphql-transformer-core@7.4.7) (2022-02-25)
+
+**Note:** Version bump only for package graphql-transformer-core
+
+
+
+
+
 ## [7.4.6](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformer-core@7.4.5...graphql-transformer-core@7.4.6) (2022-02-15)
 
 **Note:** Version bump only for package graphql-transformer-core